Have DOS_MultiplexFunctions ax = 0x1216 use a seperate fake SFT table
This interrupt is a bit hack-ish in the way it works.
It writes out the SFT entry on demand as it is called and returns a pointer.
This is because DOSBox does not actually write SFT entries to DOS memory on normal file operations.

We were running into problems with this interrupt stepping on memory used for other DOS data structures.
Avoid this problem by giving this interrupt its own dedicated memory.