Revert "Move System File Table memory location"
This reverts commit b31a7fdc1273b181822a1d163ab9bd962d365875.

This was the previous fix to the DOS_MultiplexFunctions ax = 0x1216 problem.
However, it caused a regression in that DOS programs run inside Windows 3.1 would crash sometimes triggering a segmentation fault of the entire emulator.

Windows 3.1 expects the SFT to be located at a fixed position in DOS memory.
The multiplex function now uses its own memory to avoid stepping on this "real" SFT table.