Use GUS's AdLib command reg mirroring function in OPL
---

Thanks to GitHub user @gmcn42 for the following information:

I just ran into this problem in dosbox-x 0.84.2 and found
this issue while googling for a solution. I know, different
fork and all but I guess the GUS code might still be similar.

I did some digging into GUS docs, decompiled the ULTRINIT
binary and used the dosbox debugger on it. The gist: It's a
port conflict between the GUS and the OPL at 388h. Real cards
did have a similar issue but this could be cured by running
ULTRINIT which would succeed and (probably) deactivate the
port on the GUS if SBOS won't be used. I can confirm this
works on a real GUS classic in tandem with an ESS1688 (with
OPL-compatible ESFM at 388h).

See this VOGONS thread for more discussions about it on real
hw, apparently sometimes one needs to run ULTRINIT again
after using a GUS game to get the OPL back but that's it.
What I gathered about what is going wrong in dosbox from
decompilation and debugging:

At some point, ULTRINIT 2.28 will try to write 0xAA to 388h
(hard-coded) and read it back from GUS_BASE_ADDRESS+0x0A
(24Ah in my case). GF1 seems to mirror what is written to
388h at 24Ah

If that happens, ULTRINIT is happy and goes on, if not, it
will fail at that point On real HW this seems to work with a
real OPL sharing the port because the write just goes to both
devices, reading back does not conflict because port 24Ah
belongs to GUS alone

+ This looks to me like it was intentional by Gravis due to
  the conflict-robust mirroring

+ Afterwards ULTRINIT seems to deactivate the 388h port on
  the GUS, SBOS will reactivate if loaded

In dosbox the write to 388h seems to not reach the GUS if an
OPL is present as well, the value read from 24Ah is 0x04
instead of 0xAA => ULTRINIT fails