Do not re-use file handles in DOS_FCBOpen
This commit has been in the tree for over 20 years but it is wrong.
We need to always allocate a new file handle so they don't share file positions, etc with other open handles referencing the same file.

FCB (File Control Block) is a legacy method of opening files.
The FCB block is stored entirely in the program's memory.
In DOS 2.0, this was deprecated in favor of file handles pointing to DOS system memory.