Render the audio stream in "no sound" mode
This ensures we internally generate the audio stream in "no sound" mode
too, so it can be captured and the "rendered frames" counters keep
progressing.

Prior to this change, certain programs that rely on these counters for
timing were stalled in "no sound" mode (e.g., Fast Tracker II and PT4DOS).