Fix audio stretching regression of `MixerChannel::Mix`
Fixes regression introduced in 4c399890af6

This regression is hard to test. One reliable test case is the
"Conquests of the Longbow: The Legend of Robin Hood" demo. At the start
of the demo, there is a digitised speech sample that goes "Robin Hood,
your bow is needed!"

Without this fix, the "Robin Hood" part is almost completely cut off.