Fix `INT10_ReloadFont()` regression & refactor
This fixes a regression in the "Druid: Daemons of the Mind" installer
where the whitespaces of the custom VGA font appear corrupted
(introduced in 5934a75).

`INT10_ReloadFont()` should trigger the loading of alternate characters
for 9-dot VGA fonts only, while `INT10_LoadFont()` should never do this.
This has regressed during a recent cleanup & refactoring job.

This change also implements this conditional alternate charset loading
mechanism more cleanly. Previously, the `font_block` argument was used
as a "side channel" for signalling this criteria via a hack (by using a
magic value of 0x80 which is not a valid value in the real BIOS call).