Upgrade GUS's mix control and address select registers with bit_view
The named bits helped clarify that one of the variables in the IRQ
handler was wrong. It was named "lines" but the bit it was checking
(0x8) is actually the mix control's latch bit.

The address select register also made is easy to process and log the
secondary IRQ and DMA address in the 2xB IO write call. These
secondary addresses aren't really used, but we might as well provide
that scaffolding if we want to handle them in the future.