Upgrade the GUS DMA control register with bit_view
This eliminates the previous hack that used a 16-bit register to store
the pending DMA terminal count IRQ status in bit 9, which the
documentation doesn't describe (the docs say it's an 8-bit register).

The previous code used bit twiddling to manage this divergence from the
docs - so bit 9 had to be shifted, masked, and/or inverted when the DMA
control register was written or read.