Move `ListDevices()` out of the `MidiDevice` interface
If the `ListDevices()` functions of the MIDI devices can be invoked
without an instance of that particular MIDI device existing, we can
simplify the whole MIDI device handling to follow simple RAII patterns.

There will be no need for all MIDI devices to be instantiated, even
though only one is active, just to be able to call `ListDevices()` on
them.

Incidentally, this will also allow us to introduce a second MPU-401
interface much more easier in the future if needed.