Eliminate the internal `InstantLaunch` verbosity level
The InstantLaunch verbosity level isn't available to users,
and (in addition to being a verbosity level), it seemed to
have a second purpose to indicate if the user was launching
a DOS executable passed into the DOSBox command line.

This commit drops this verbosity level (it's not needed),
and fixes a problem where the `quiet` verbosity level was
incorrectly used when it should be `low` per the
`startup_verbosity` conf documentation.