Directly process the Sound Blaster's DAC samples
Uses a simple DAC class to measure and run the Sound Blaster's
audio channel at the DAC data rate, processing samples on a
per-frame basis.

This eliminates the mixer's "Add Stretched" and the
AudioVector's similar stretch-vector derived class, both of
which were only used by the Sound Blaster DAC.

Because the DAC uses the per-frame callback timing, this
commit also ensure we continue to use the per-tick callback
timing for DMA playback (at the point when the transfer is
unmasked).