Fix GUS resampling regression & emulate the GUS' output more authentically
Previously, the output was rendered at the GUS' internal mixer rate
which is dependent on the number of active voices, then this output was
LERP-upsampled to the host mixer rate, so typically to 44.1 or 48 kHz.

Typical GUS mixer rate examples:

- Star Control II: 28 active voices, 22050 Hz mixer sample rate

- DOOM I & II: 28 active voices, 22050 Hz mixer sample rate (with the
  maximum 8 sfx channels configured)

- Pinball Fantasies: 16 active voices, 38587 Hz mixer sample rate

This LERP upsampling basically had a weird smoothing/low-pass filtering
effect on the output which was not too far off from the GUS Classic
sound by fluke.

By mistake, the LERP upsampler was the default upsampler for all
channels that did not specifically set the resampling method. This was
fixed in commit 5abef42, but that removed the "beneficial"
high-end-smoothing effect of the LERP upsampler on the GUS's output; its
sound became subtly different in the high frequencies (subjectively, the
top end became a bit harsher).

This PR introduces a better approach that gets us much closer to the
actual GUS Classic sound:

- On real hardware, the GUS always outputs a 44.1 kHz sample stream to
  the DAC; the channel count dependent mixer rate degradation is
  achieved via dropping samples in the GUS mixer.

- Our GUS emulation renders at the mixer rate, but with ZOH upsampling
  to 44.1 kHz we can approximate this sample dropping and fixed 44.1
  kHz output rate behaviour almost perfectly. We then resample this
  with Speex to the host rate which is usually 48 kHz.

- But this adds a lot of high-end, so we need to faithfully emulate the
  analog low-pass filter stage of the GUS as well. On real hardware,
  some gentle low-pass filtering starts from about 7-9 kHz; at 10kHz the
  drop is about -3 dB, and at 20kHz around -9 dB. A 6 dB/oct LPF from 8
  kHz gets us very close to this sound (this has been tweaked by
  comparing the filtered output to real hardware GUS Classic
  recordings).