Fix edge case where CDROM mixer hangs at the end of a track
When the track got to the end, it was doing some back-and-forth
logic going between time and sector number to figure out where to seek next.

It can hang if either the conversion is slightly off
or the MP3 decoder returns slightly fewer frames that it was expecting.

Even on MP3 files where it didn't hang, it sometimes did 10+ seeks
back-to-back each seeking to the last second of the file.

New code skips all that conversion buisness and just goes to the start of the next track :)