Remove `auto` option from the `vsync` setting
The removed code attempted to infer whether enabling vsync is desirable
or the best thing to do one any given setup. This is simply impossible
to do with a reasonable degree of reliability on the three platforms we
support (Windows, macOS, and Linux). There are lots of edge cases,
platform specific quirks, driver quirks, and GPU-specific behaviour
involved, not to even mention customised driver settings.

Simply put, this automatic behaviour was never working great; it was
full of unexplained magic heuristics, and it added significant
complexity and maintenance burden to the code.