Fix sign conversion warnings
Commit 80981c965a621a48c73c5cf6c58cae699fa1e4f7 changed types
of width and height of ImageInfo from uint16_t to int

This introduced these sign conversion warnings in ReelMagic's video_mixer.cpp

It's not so easy to change the RealMagic variable types
as there is bitshifting and pointer arithmatic involved.

Adding check_casts's to supress the warnings for now
and trigger an assert if the values are out of bounds.