Remove sRGB texture & framebuffer support
We only started using sRGB textures & framebuffers to make the `sharp`
shader faster on the Pi -- that was the only reason. The CRT shaders
don't rely on hardware-assisted sRGB decoding/encoding; they do their
own multisampling and gamma encoding/decoding via `pow()` functions in
the shader code anyway.

With this change, the `sharp` and `catmull-rom` shaders do the same
(which are more lightweight than `crt-hyllian` anyway).

On desktop CPUs the difference between hardware sRGB decoding/encoding
and doing it in the shader is extremely minuscule; the CPU is always the
bottleneck with these shaders.

Here's a very detailed article describing the murky hardware sRGB
texture & FBO support:

https://unlimited3d.wordpress.com/2020/01/08/srgb-color-space-in-opengl/